ganti tipe data primary key

+++ src/test/sql/skema.sql
@@ -1,5 +1,5 @@
 create table enable_fitur (
-  id INT PRIMARY KEY AUTO_INCREMENT,
+  id BIGINT PRIMARY KEY AUTO_INCREMENT,
   mhsw_id VARCHAR(10) NOT NULL,
   fitur VARCHAR(20) NOT NULL,
   enable BOOLEAN NOT NULL
